+++ CHANGELOG.md
@@ -1,1 +1,10 @@
 # Changelog **TODOvue Theme Button**
+
+## v1.0.0
+### Dependencies Update
+* Update `todovue-css` to version 1.8.0
+* Update `todovue-demo`to version 2.3.0
+### Features
+* Create `todovue-theme-button` component
+* Add `todovue-theme-button` package to NPM
+* Unit tests are added